requirements.txt
amqp==1.4.9
anyjson==0.3.3
arrow==0.12.1
billiard==3.3.0.23
celery==3.1.23
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-filter==1.1.0
djangorestframework==3.8.2
future==0.16.0
idna==2.6
Jinja2==2.8
jinja2-time==0.1.0
kik==1.0.7
kombu==3.0.37
Markdown==2.6.11
MarkupSafe==1.0
messengerbot==0.1.4
mimesis==2.0.1
psycopg2-binary==2.7.4
pycrypto==2.6.1
python-dateutil==2.7.2
python-telegram-bot==4.2.0
pytz==2018.3
requests==2.10.0
shortuuid==0.4.3
six==1.10.0
urllib3==1.22
uWSGI==2.0.17
whitenoise==3.3.1
